requirements.txt
wheel
azure-functions
azure-storage-blob
azure-servicebus
azure-storage-queue